scripts/requirements.txt
anthropic>=0.18.0
pyyaml>=6.0

# Optional: For favicon generation (only if using Python script)
# pillow>=10.0.0
# cairosvg>=2.7.0
